requirements.txt
pillow
matplotlib
numpy
scikit-image
torch
torchvision